Enable telemetry and crash reports by default
Remove the onboarding diagnostics step and the one-time migration
sheet. Both data streams now default to on (nil flag = enabled);
opt-out remains in Settings > Features > App Diagnostics. Update
privacy policy, drop obsolete shadow-mode PRIVACY.md.
