+++ Trio/Sources/Modules/Telemetry/View/TelemetryPrivacyView.swift
@@ -26,7 +26,7 @@ struct TelemetryPrivacyView: View {
                     Text("What stays on your device")
                         .font(.headline)
                     Text(
-                        "All glucose, insulin, and carb data. All therapy settings (basal, ISF, carb ratio, glucose targets). Your Nightscout URL and API token. Your Tidepool credentials. Remote-command secrets and APNS keys. Time zone. Location data. Logs are never sent automatically; sharing them remains a user-initiated flow."
+                        "All glucose, insulin, and carb data. All therapy settings (basal, ISF, carb ratio, glucose targets). Your Nightscout URL and API token. Your Tidepool credentials. Remote-command secrets and APNS keys. Location data. Logs are never sent automatically; sharing them remains a user-initiated flow."
                     )
                 }
 
@@ -42,7 +42,7 @@ struct TelemetryPrivacyView: View {
                     Text("Opt out")
                         .font(.headline)
                     Text(
-                        "Use the diagnostics-sharing chooser above. Pick \"Crash Reports Only\" to keep crash reporting but disable telemetry, or \"Disable Sharing\" to turn off both. Changes take effect immediately."
+                        "Crash reporting and anonymous usage telemetry are enabled by default. Use the diagnostics-sharing chooser above to opt out: pick \"Crash Reports Only\" to keep crash reporting but disable telemetry, or \"Disable Sharing\" to turn off both. Changes take effect immediately."
                     )
                 }
             }
